MB-67520: Fix many narrowing warnings in dcp
DCP code is mostly mixing u32/u64 around flow control, the only part
of flow control that is limited to u32 is the protocol message for
acking bytes (defines a u32 field). Internally we should use size_t
throughout, e.g. getMessageSize returns the size of the message and
that doesn't have to be limited to u32. Flow control counters are
already 64 (now size_t) and we could of in theory accumulated more
data in the consumer that we can ack in one message - this "overflow"
is now checked against in flow-control.cc, but the ack trigger
boundary is far far lower than 4billion.

Reviewed-on: https://review.couchbase.org/c/kv_engine/+/230725
Tested-by: Build Bot <build@couchbase.com>

+++ engines/ep/src/dcp/backfill-manager.cc
@@ -425,7 +425,7 @@ void BackfillManager::moveSnoozingToActiveQueue() {
     }
 }
 
-int BackfillManager::getNumInProgressBackfills(
+size_t BackfillManager::getNumInProgressBackfills(
         const std::unique_lock<std::mutex>& lh) const {
     return initializingBackfills.size() + activeBackfills.size() +
            snoozingBackfills.size() + numInProgressUntrackedBackfills;
